Fix accidental regressions from the float16/bfloat16 type refactor
- src/EmbeddingSpMDMAutovec.cc: remove a stray #ifdef __linux__ / #endif
  that wrapped the entire translation unit, compiling out every autovec
  CPU kernel (EmbeddingSpMDM8Bit_autovec, EmbeddingSpMDMNBit_autovec,
  EmbeddingSpMDM_autovec, EmbeddingSpMDMRowWiseSparse_autovec,
  EmbeddingSpMDMFP8_autovec, and all explicit instantiations) on any
  non-Linux target. Restore the now-needed EmbeddingSpMDMPrefetch.h and
  fbgemm/Utils.h includes the L1/L2 prefetch helpers depend on, and
  restore the constexpr on stride_SpMDMWithStrides/stride_SpMDMNBitWith
  that got dropped along the way.
- src/EmbeddingSpMDMSve.h: EmbeddingSpMDM8Bit_Sve_Fp16 and
  EmbeddingSpMDMNBit_Sve_Fp16 still gated on
  `std::is_same_v<OutType, uint16_t>`, a check left over from when fp16
  output was represented as a raw uint16_t. Both call sites now
  instantiate these with OutType = fbgemm::float16, so the check was
  always false and the SVE fp16-accumulator fast path
  (FBGEMM_TBE_SVE_FP16_ACCUMULATOR) always returned false instead of
  computing a result. Fix the check to compare against float16.

+++ src/EmbeddingSpMDMSve.h
@@ -813,9 +813,9 @@ bool EmbeddingSpMDM8Bit_Sve_Fp16(
     const int64_t output_stride,
     const int64_t input_stride,
     const bool scale_bias_last) {
-  // This kernel is only dispatched for fp16 output (OutType == uint16_t).
+  // This kernel is only dispatched for fp16 output (OutType == float16).
   // All paths produce fp16 directly — no fp32 widening.
-  if constexpr (!std::is_same_v<OutType, uint16_t>) {
+  if constexpr (!std::is_same_v<OutType, float16>) {
     return false;
   }
 
@@ -1823,9 +1823,9 @@ bool EmbeddingSpMDMNBit_Sve_Fp16(
     const int64_t output_stride,
     const int64_t input_stride,
     const bool scale_bias_last) {
-  // This kernel is only dispatched for fp16 output (OutType == uint16_t).
+  // This kernel is only dispatched for fp16 output (OutType == float16).
   // All paths produce fp16 directly — no fp32 widening.
-  if constexpr (!std::is_same_v<OutType, uint16_t>) {
+  if constexpr (!std::is_same_v<OutType, float16>) {
     return false;
   }
